+++ pylinkage/collections/__init__.py
@@ -1,0 +1,2 @@
+from .agent import Agent
+from .mutable_agent import MutableAgent