+++ pylinkage/collections/agent.py
@@ -1,0 +1,10 @@
+from collections import namedtuple
+
+
+class Agent(namedtuple("Agent", ["score", "dimensions", "init_positions"])):
+    """
+    A class that uniformizes a linkage optimization.
+
+    It is roughly a namedtuple with preassigned fields.
+    """
+    pass