+++ pylinkage/interface/__init__.py
@@ -1,0 +1,14 @@
+"""
+Classes that represent the main interface with the library.
+"""
+from .exceptions import (
+    UnbuildableError,
+    HypostaticError
+)
+from .joint import (
+    Static,
+    Pivot,
+    Crank,
+    Fixed
+)
+from .linkage import Linkage