+++ tests/test_linkage.py
@@ -1,66 +1,35 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
-Created on Fri Apr 16 17:19:47 2021
-
-@author: HugoFara
+Test cases for a Linkage.
 """
-
 import unittest
-from math import sqrt
-from pylinkage.exceptions import UnbuildableError
-from pylinkage.linkage import Pivot, Fixed
+import pylinkage as pl
 
 
-class TestPivot(unittest.TestCase):
-    """Test Pivot Joint."""
+class TestLinkage(unittest.TestCase):
+    def __init__(self, method_name):
+        super().__init__(method_name)
+        # Main motor
+        self.crank = pl.Crank(
+            0, 1,
+            joint0=(0, 0),  # Fixed to a single point in space
+            angle=0.31, distance=1,
+            name="B"
+        )
+        # Close the loop
+        self.pin = pl.Pivot(
+            3, 2,
+            joint0=self.crank, joint1=(3, 0),
+            distance0=3, distance1=1, name="C"
+        )
 
-    pivot1 = Pivot(0, 0)
-
-    def test_buildable(self):
-        """Upper intersect test."""
-        pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
-        pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, sqrt(.75)))
-
-    def test_under_intersect(self):
-        """Under intersect test."""
-        pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(y=-1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
-        pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, -sqrt(.75)))
-
-    def test_limit_intersect(self):
-        """Test system almost breaking."""
-        pivot2 = Pivot(2, 0)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
-        pivot3.reload()
-        self.assertEqual(pivot3.coord(), (1, 0))
-
-    def test_no_intersect(self):
-        """Test system almost breaking."""
-        pivot2 = Pivot(0, 3)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
-        with self.assertRaises(UnbuildableError):
-            pivot3.reload()
-
-
-class TestFixed(unittest.TestCase):
-    """Test Fixed_Joint."""
-
-    pivot1 = Pivot(0, 0)
-
-    def test_pos(self):
-        """Test Fixed_Joint positioning."""
-        pivot2 = Pivot(1, 0)
-        fixed = Fixed(joint0=self.pivot1, joint1=pivot2, angle=0, distance=1)
-        fixed.reload()
-        self.assertEqual(fixed.coord(), (1, 0))
+    def test_definition(self):
+        # Linkage definition
+        my_linkage = pl.Linkage(
+            joints=[self.crank, self.pin],
+            order=[self.crank, self.pin],
+            name="My four-bar linkage"
+        )
+        self.assertTupleEqual(my_linkage.joints, (self.crank, self.pin))
 
 
 if __name__ == '__main__':